pyproject.toml
[tool.poetry]
name = "file-convection"
version = "1.0.0"
description = "File conversion application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.13.7,<3.14.0"
Flask = "^3.1.2"
Werkzeug = "^3.1.3"
waitress = "^3.0.0"
numpy = "^2.3.4"  # Updated for Python 3.13 compatibility
pandas = "^2.3.2"
pdfplumber = "^0.11.7"
python-dotenv = "^0.19.2"
tabula-py = "^2.10.0"
JPype1 = ">=1.4.2,<2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"